moves helper methods that adds local variables to CecilifierContextExtensions
